Was able to reproduce issue #2846 consistently by having the dbcontext open / edited (not saved) and invoking reverse engineer.  System dialog would open which corrupted the existing process flow.   Removed code in StatusBarHelper that I added early in development that didn't allow system dialog to place nicely.  Also refactored flow from page 3 to page 4 for a more consistent user experience.   Tested by opening all dialogs and taking steps noted above to reproduce - had expected results with no issues.
